+++ source/CCustomOpcodeSystem.h
@@ -6,6 +6,7 @@
 
 namespace CLEO
 {
+    const size_t MAX_STR_LEN = 0xff; // max length of string type parameter
     enum OpcodeResult : char
     {
         OR_CONTINUE = 0,